// src/types/types.ts
export interface Product {
  id: number;
  title: string;
  price: number;
  description: string;
  category: string;
  image: string;
  rating: {
    rate: number;
    count: number;
  };
}

export interface CartItem extends Product {
  quantity: number;
}

import { store } from '../main';
export type RootState = ReturnType<typeof store.getState>;
export type AppDispatch = typeof store.dispatch;